完善Server消息推送

+++ CNSD2/Server/Server.h
@@ -23,7 +23,7 @@ private slots:
 	void getMsg(const QString msg, unsigned int id);
 	void getData(const std::string data, unsigned int id);
 	void getConnection();
-	void cilentDisconnected(const unsigned short id);
+	void cilentDisconnected(const unsigned int id);
 
 	void socketHandleThreadStopped(unsigned int id);
